refactor(sdk): Convert webpack.embedding-sdk.config to rspack and prepare for Hosted Bundle
