+++ .eslintrc.js
@@ -1,6 +1,8 @@
 /* eslint-disable import/no-commonjs */
 /* eslint-disable no-undef */
 
+const path = require("path");
+
 // `postcss-modules` lints css modules class names, but it currently crashes
 // eslint on vscode. If you use webstorm or want to run the lint for the cli, you
 // can use this flag to enable it. This is set to true in CI
@@ -155,6 +157,7 @@ module.exports = {
     "import/internal-regex": "^metabase/|^metabase-lib/",
     "import/resolver": {
       webpack: {
+        config: path.resolve(__dirname, "./rspack.main.config.js"),
         typescript: true,
       },
     },
@@ -203,8 +206,9 @@ module.exports = {
         "e2e/**/*",
         "**/tests/*",
         "release/**/*",
-        "webpack.config.js",
         "rspack.config.js",
+        "rspack.main.config.js",
+        "rspack.embedding-sdk-bundle.config.js",
       ],
       rules: {
         "no-color-literals": "off",
@@ -221,8 +225,7 @@ module.exports = {
         "e2e/**/*",
         "**/tests/*",
         "release/**/*",
-        "webpack.config.js",
-        "rspack.config.js",
+        "rspack.main.config.js",
       ],
       rules: {
         "i18next/no-literal-string": "off",
@@ -331,6 +334,12 @@ module.exports = {
         "import/order": "off",
         "import/no-unresolved": "off",
         "no-color-literals": "off",
+      },
+    },
+    {
+      files: ["frontend/build/**/*.js"],
+      rules: {
+        "import/no-commonjs": "off",
       },
     },
   ],
