+++ .github/file-paths.yaml
@@ -29,8 +29,8 @@ frontend_configs: &frontend_configs
   - ".eslintrc.js"
   - "postcss.config.js"
   - "rspack.config.js"
-  - "webpack.config.js"
-  - "webpack.embedding-sdk.config.js"
+  - "rspack.main.config.js"
+  - "rspack.embedding-sdk-bundle.config.js"
   - "rspack.static-viz.config.js"
   - ".nvmrc"
