+++ .storybook/main.ts
@@ -1,7 +1,9 @@
 import type { StorybookConfig } from "@storybook/react-webpack5";
-const appConfig = require("../webpack.config");
+const appConfig = require("../rspack.main.config.js");
 const webpack = require("webpack");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
+const { CSS_CONFIG } = require("../frontend/build/shared/rspack/css-config");
 
 const mainAppStories = [
   "../frontend/**/*.mdx",
@@ -57,9 +59,19 @@ const config: StorybookConfig = {
           ...(config.module?.rules ?? []).filter(
             (rule) => !isCSSRule(rule) && !isSvgRule(rule),
           ),
-          ...appConfig.module.rules.filter(
-            (rule) => isCSSRule(rule) || isSvgRule(rule),
-          ),
+          ...appConfig.module.rules.filter((rule: any) => isSvgRule(rule)),
+          // We use MiniCssExtractPlugin, because Storybook can't properly work with `rspack.CssExtractRspackPlugin`
+          {
+            test: /\.css$/,
+            use: [
+              {
+                loader: MiniCssExtractPlugin.loader,
+                options: { publicPath: "./" },
+              },
+              { loader: "css-loader", options: CSS_CONFIG },
+              { loader: "postcss-loader" },
+            ],
+          },
         ],
       },
     };
@@ -67,5 +79,5 @@ const config: StorybookConfig = {
 };
 export default config;
 
-const isCSSRule = (rule) => rule.test?.toString() === "/\\.css$/";
-const isSvgRule = (rule) => rule.test?.test(".svg");
+const isCSSRule = (rule: any) => rule.test?.toString() === "/\\.css$/";
+const isSvgRule = (rule: any) => rule.test?.test(".svg");
