+++ enterprise/frontend/src/embedding-sdk/.eslintrc.js
@@ -7,7 +7,7 @@ module.exports = {
       webpack: {
         config: path.resolve(
           __dirname,
-          "../../../../webpack.embedding-sdk.config.js",
+          "../../../../rspack.embedding-sdk-bundle.config.js",
         ),
         typescript: true,
       },
