+++ frontend/src/metabase/app.js
@@ -6,6 +6,7 @@ import "regenerator-runtime/runtime";
 // This is conditionally aliased in the webpack config.
 // If EE isn't enabled, it loads an empty file.
 // Should be imported before any other metabase import
+// eslint-disable-next-line import/no-duplicates
 import "ee-overrides";
 
 import "metabase/lib/dayjs";
@@ -25,6 +26,7 @@ import "metabase/plugins/builtin";
 
 // This is conditionally aliased in the webpack config.
 // If EE isn't enabled, it loads an empty file.
+// eslint-disable-next-line import/no-duplicates
 import "ee-plugins";
 
 // Set nonce for mantine v6 deps
