+++ package.json
@@ -334,7 +334,6 @@
     "stylelint-config-css-modules": "^4.4.0",
     "stylelint-config-standard": "^36.0.0",
     "swc-loader": "^0.2.6",
-    "terser-webpack-plugin": "^5.3.9",
     "ts-morph": "^25.0.0",
     "tsx": "^4.19.3",
     "typedoc": "0.28.5",
@@ -381,7 +380,7 @@
     "mage": "./bin/mage",
     "build": "yarn build:cljs && yarn build:js",
     "build-embedding-sdk": "yarn && yarn build-release:cljs && yarn embedding-sdk:generate-package && yarn embedding-sdk:generate-nextjs-compat && yarn build-release:sdk && yarn build-embedding-sdk-cli && yarn embedding-sdk:fixup-types-imports",
-    "build-embedding-sdk:pure": "WEBPACK_BUNDLE=development webpack --config webpack.embedding-sdk.config.js",
+    "build-embedding-sdk:pure": "WEBPACK_BUNDLE=development rspack --config rspack.embedding-sdk-bundle.config.js",
     "build-embedding-sdk:watch": "yarn embedding-sdk:generate-package && yarn embedding-sdk:generate-nextjs-compat && yarn build-embedding-sdk:pure --watch",
     "build-embedding-sdk-cli": "rspack --config rspack.embedding-sdk-cli.config.js",
     "build-embedding-sdk-cli:watch": "yarn build-embedding-sdk-cli --watch",
@@ -394,13 +393,13 @@
     "build-release": "yarn build-release:cljs && yarn build-release:js",
     "build-release:cljs": "yarn clean-release:cljs && shadow-cljs release app",
     "build-release:js": "WEBPACK_BUNDLE=production rspack build",
-    "build-release:sdk": "WEBPACK_BUNDLE=production webpack --config webpack.embedding-sdk.config.js",
+    "build-release:sdk": "WEBPACK_BUNDLE=production rspack --config rspack.embedding-sdk-bundle.config.js",
     "build-release:static-viz": "WEBPACK_BUNDLE=production rspack --config rspack.static-viz.config.js",
     "build-shared": "yarn && rspack --config rspack.shared.config.js",
     "build-static-viz": "yarn && yarn build-release:cljs && WEBPACK_BUNDLE=production rspack --config rspack.static-viz.config.js",
     "build-static-viz:watch": "rspack --config rspack.static-viz.config.js --watch",
     "build-static-viz:watch-wait": "yarn wait:cljs; yarn build-static-viz:watch",
-    "build-stats": "yarn && webpack --json > stats.json",
+    "build-stats": "yarn && rspack --json > stats.json",
     "build-storybook": "NODE_OPTIONS=--max-old-space-size=8192 storybook build",
     "build-storybook-embedding-sdk": "NODE_OPTIONS=--max-old-space-size=8192 IS_EMBEDDING_SDK=true storybook build -c .storybook-sdk --quiet",
     "build-watch": "yarn concurrently -n 'cljs,js' 'yarn build-watch:cljs' 'yarn build-watch:js'",
